+++ .github/workflows/generate_java_sources.yml
@@ -34,8 +34,8 @@ jobs:
     
       - name: Generate SDKs
         run: |
-          chmod +x ./scripts/build_java_sdk.sh
-          ./scripts/build_java_sdk.sh
+          chmod +x ./scripts_legacy/build_java_sdk.sh
+          ./scripts_legacy/build_java_sdk.sh
         shell: bash
 
       - name: Test SDKs
