+++ .github/workflows/generate_php_sources.yml
@@ -49,8 +49,8 @@ jobs:
 
       - name: Generate SDKs
         run: |
-          chmod +x ./scripts/build_php_sdk.sh
-          ./scripts/build_php_sdk.sh
+          chmod +x ./scripts_legacy/build_php_sdk.sh
+          ./scripts_legacy/build_php_sdk.sh
         shell: bash
 
       - name: Test SDKs
