+++ .github/workflows/generate_python_sources.yml
@@ -35,8 +35,8 @@ jobs:
 
       - name: Generate SDKs
         run: |
-          chmod +x ./scripts/build_python_sdk.sh
-          ./scripts/build_python_sdk.sh
+          chmod +x ./scripts_legacy/build_python_sdk.sh
+          ./scripts_legacy/build_python_sdk.sh
         shell: bash
 
       - name: Test SDKs
