--- scripts/git-track/java.txt
@@ -1,11 +1,0 @@
-src/
-docs/
-*.md
-*.gradle
-.gitignore
-.openapi-generator/
-gradlew
-gradlew.bat
-gradle/
-pom.xml
-LICENSE.txt
