--- scripts/git-track/php.txt
@@ -1,9 +1,0 @@
-lib/
-docs/
-examples/
-test/
-composer.json
-*.md
-*.txt
-.openapi-generator/
-*.xml.dist
