--- scripts/git-track/python.txt
@@ -1,10 +1,0 @@
-criteo_marketing/
-docs/
-examples/
-test/
-*.py
-*.md
-*.txt
-.gitignore
-.openapi-generator/
-tox.ini
