+++ scripts/shared/utils.py
