+++ scripts/test_sdk.py
@@ -1,13 +1,13 @@
 import sys, getopt
 from os import path
 
-from push_sdk.clients.fs_client import FsClient
-from push_sdk.clients.os_client import OsClient
-from push_sdk.utils import get_logger
+from .shared.clients.fs_client import FsClient
+from .shared.clients.os_client import OsClient
+from .shared.utils import get_logger
 
-from push_sdk.java_sdk_test_action import JavaSdkTestAction
-from push_sdk.python_sdk_test_action import PythonSdkTestAction
-from push_sdk.php_sdk_test_action import PhpSdkTestAction
+from .languages.java.java_sdk_test_action import JavaSdkTestAction
+from .languages.python.python_sdk_test_action import PythonSdkTestAction
+from .languages.php.php_sdk_test_action import PhpSdkTestAction
 
 logger = get_logger()
