+++ scripts/tests/dummies/dummy_git_client.py
@@ -1,5 +1,5 @@
 from ..result_or_exception import ResultOrException
-from ...clients.git_client import IGitClient, GitException
+from ...shared.clients.git_client import IGitClient, GitException
 
 class DummyGitClient(IGitClient):
     def __init__(self):
