+++ scripts/tests/dummies/dummy_os_client.py
@@ -1,4 +1,4 @@
-from ...clients.os_client import IOsClient
+from ...shared.clients.os_client import IOsClient
 from ..result_or_exception import ResultOrException
 
 class DummyOsClient(IOsClient):
