+++ scripts/tests/result_or_exception.py
