+++ scripts_legacy/build_php_sdk.sh
